1.5 Update
Fixed tab movement.  This now works again.  Issue was caused because
events were propagated across all tabs.

Fixed extension not being loaded in new tabs.  This was caused because
messages from the global page can only be sent to the 'active' tab (not
the case when a new tab is opened).  The extension is now
enabled/disabled on tab activation.

Added new feature.  CTRL-SHIFT-F now opens links in a new tab!!