Edit mousetrap to restore previous ctrl behavior
Prior to integrating mousetrap, ctrl commands were always available,
ex. you could ctrl-f in an input. Since mousetrap disables itself for
inputs, this behavior was regressed.

This patches mousetrap to allow the event to trigger in an input
if it has a modifier key. If mousetrap sticks around in this codebase,
a more generic solution should be attempted to be pushed upstream to
mousetrap.