Fix some odd edge case behavior
At some point I determined keypress and keydown didn't work, and needed
to hook into keyup. This is not true anymore, and this patch makes
things work smoothly.

Note that keypress ignores meta keys, which breaks the ability to not
have a modifier prefix for commands. Thus, we do need to use keydown.