Detection improvements: Add detection by cursor style.
For custom widgets that don't use form controls or ARIA markup,
we can still detect clickable elements by looking for items
having cursor style set to "pointer".

As this property is usually inherited, we look only for elements
whose parent doesn't have that cursor style too.

Usually, normal elements and ARIA roles are enough, and this kind of
detection introduces too much noise. But for some pages, having this
capability proves itself very useful. Therefore, we introduce a setting
to turn on/off this kind of detection. By default, it's set to off.