Bootstrap Safari App Extension project
This change is just the result of generating a new project with a Safari
Extension target + the vimari files added. It builds but it doesn't work
in the browser.

To be able to test this you need Safari 12, I think. So I downloaded the
Safari Technology Preview to be able to run this extension. Also you
need Xcode - use the latest version to be on the safe side. And if
you're not in the Apple Developer Program you need to Enable Unsigned
Extensions, you can google for how to do that.

To run it in the browser:
  1) Have an instance of Safari Technology Preview open
  2) Open the project with Xcode
  3) Run the project. A small window will open but ignore it.
  4) Go to Safari -> open settings (CMD + ,) -> Enable extension under
     "Extensions"
  5) Profit!

To debug:
  * To make sure the extension is loading you can inspect the "Resources"
    tab in your development tools and you should see "Extension Scripts"
    being loaded. From here you can see the JS code and set up breakpoints
    etc.
  * Use console.log(), as you normally would
  * Do small changes to the JS because if for some reason something is
    wrong with the JS code, Xcode will simply not send it to the browser
    and it won't tell you why. So do small changes, test if they worked,
    and then proceed.

I added this information because I found the information Apple give out
too thin - I had to figure out most of this on my own.