Fix inability to switch in pinned tabs
Due to not being able to request the containing tab of a pinned tab (as
they don't have one), a bug was introduced that did not allow the user
to switch tabs if the active tab was a pinned one. This change
introduces some additional logic that in this case returns the
current active window.