Remove redundant openNewTab function
Likely because of the previous working the code for opening a new tab
was extracted into a separate function. However now that we call the
Vimari Extension for this behaviour the code should be called in the
same way as for example the close tab action. Therefore it is placed
directly in the actionmap.