Update Mousetrap to 1.6.5
Upon manual testing this did not break any functionally. Since this
updates the Minor from 3 to 6 there should not be any breaking changes.

The library is upgraded such that a minor modification can be made,
allowing us to properly isolate website behaviour from Vimari (normal
mode). I figured it best to perform this modification on the latest
version of Mousetrap such that possible new features are included.