Use window.open for opening a new tab
This replaces the usage of the extension for this task, which removes
the ability to open a new tab without making it active. However for the
'openTab' functionality this functionality does not seem appropriate
anyway.

In return for this we get the ability to open URLs that do not conform
to the requirement set by Apple's API. This includes URLs like
`topsites://` and `about:blank`.