Fix unexpected trigger for link-hints
As the link-hint matching algorithm quits eagerly, that is when there
are no other link hints matching the typed keystrokes, a user could
trigger a Vimari binding if the last link-hint-character was also a
binding.

This is possible when the linkHintsCharacters first character is also
assigned to one of the Vimari actions. This first character is used as
padding in link-hints.js:349-350 and could therefore end up at the end
of a link hint. Triggering the scenario above if the user types all
characters from the link hint while Vimari eagerly quits before the last
character is typed (because there are no other matches given all but the
last keystrokes).

This commit resolves this issue by removing the eager evaluation of link
hints, requiring the user to type all characters displayed. This
solution was chosen as it is the simplest and most intuitive for the end
user.

This does mean that the end user can be presented with a filtered set of
exactly 1 link hint. In the future we should adjust the link-hint
algorithm to allow for variably sized link-hints such that we remove
the need for padding. Therefore removing the situation where the user
ends up with exactly 1 link hint.

For reference see issue televator-apps/vimari#190 on Github.
https://github.com/televator-apps/vimari/issues/190