Change settings to override defaults
This commit introduces a method `getMergedSettings` to the
ConfigurationModelProtocol which merges the users settings with the
default settings, prioritising the settings set by the user. This allows
us to fill up missing settings (say a newly introduced binding) with a
default value. This prevents the extension from crashing when a new
setting is missing from the user-defined settings. Therefore allowing us
to gradually introduce new settings without breaking previous
configurations.

This does not touch the users configuration file so there is a potential
that this could introduce confusion to the user when a certain binding
is introduced but not displayed in their configuration. This should be
addressed in #177.

I was not able to introduce tests for the written merging code, this
because Xcode does not allow us to test Extension targets. If anyone
knows how this works please let me know.